# openapi.yaml
openapi: 3.0.0
info:
  title: Image Hosting API
  description: Hosted image API that serves multiple sizes generated from a master image.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Provide the API key as a Bearer token
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Provide the API key via the x-api-key header
security: []
paths:
  /images:
    post:
      summary: Upload a new master image
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                description:
                  type: string
                  description: Optional description for the image
                tags:
                  type: string
                  description: Comma-separated list of tags
      responses:
        "201":
          description: Image successfully uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  width:
                    type: integer
                  height:
                    type: integer
                  format:
                    type: string
                  description:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  urls:
                    type: object
                    properties:
                      original:
                        type: string
                      thumbnail:
                        type: string
                      small:
                        type: string
                      medium:
                        type: string
                      large:
                        type: string
                      custom:
                        type: string
        "400":
          description: Bad Request (no image file provided)
        "401":
          description: Unauthorized (missing or invalid API key)
        "415":
          description: Unsupported format
    get:
      summary: List all uploaded images
      parameters:
        - in: query
          name: tags
          schema:
            type: string
          description: Comma-separated list of tags to filter by (matches ANY tag)
      responses:
        "200":
          description: A list of all images and their URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    width:
                      type: integer
                    height:
                      type: integer
                    format:
                      type: string
                    description:
                      type: string
                    tags:
                      type: array
                      items:
                        type: string
                    urls:
                      type: object
                      properties:
                        original:
                          type: string
                        thumbnail:
                          type: string
                        small:
                          type: string
                        medium:
                          type: string
                        large:
                          type: string
                        custom:
                          type: string
  /all:
    get:
      summary: Alias for GET /images to list all uploaded images
      parameters:
        - in: query
          name: tags
          schema:
            type: string
          description: Comma-separated list of tags to filter by (matches ANY tag)
      responses:
        "200":
          description: A list of all images and their URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    width:
                      type: integer
                    height:
                      type: integer
                    format:
                      type: string
                    description:
                      type: string
                    tags:
                      type: array
                      items:
                        type: string
                    urls:
                      type: object
                      properties:
                        original:
                          type: string
                        thumbnail:
                          type: string
                        small:
                          type: string
                        medium:
                          type: string
                        large:
                          type: string
                        custom:
                          type: string

  /images/{id}:
    get:
      summary: Get a custom resized image
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: width
          schema:
            type: integer
        - in: query
          name: height
          schema:
            type: integer
        - in: query
          name: fit
          schema:
            type: string
            enum: [cover, contain, fill, inside, outside]
        - in: query
          name: format
          schema:
            type: string
            enum: [jpeg, jpg, png, webp, avif]
      responses:
        "200":
          description: The resized image
          content:
            image/*:
              schema:
                type: string
                format: binary
        "400":
          description: Bad Request (missing width/height or invalid params)
        "404":
          description: Image not found

    delete:
      summary: Delete an image master and cached variants
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  id:
                    type: string
        "401":
          description: Unauthorized
        "404":
          description: Image not found

  /images/{id}/original:
    get:
      summary: Get the untouched master image
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The original image
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Image not found

  /images/{id}/{preset}:
    get:
      summary: Get a preset size of the image
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: preset
          required: true
          schema:
            type: string
            enum: [thumbnail, small, medium, large]
        - in: query
          name: format
          schema:
            type: string
            enum: [jpeg, jpg, png, webp, avif]
      responses:
        "200":
          description: The resized preset image
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Image or preset not found

  /playlist.m3u:
    get:
      summary: Get an M3U playlist of all images
      responses:
        "200":
          description: M3U playlist file
          content:
            audio/x-mpegurl:
              schema:
                type: string